Changes To tests for jenkins compatability
Had to change some tests to expect fail because of the setup of the test user. We dont have it set up for authentications of different types on the fly like other languages so we need to adapt our tests so that some tests prove a call works while expecting a failure. This is essential because our testfixture appears to make a double authentication for our test users, but since all the previously written tests rely on the authenticateUniversal, this initial authenticate conflicts with playerIds being sent as well as matchIds.